Some clippy warnings

+++ bee-pow/tests/score.rs
@@ -14,17 +14,17 @@ fn score() {
 
     let message: [u8; 8] = [0, 0, 0, 0, 0, 0, 0, 0];
 
-    assert_eq!(compute_pow_score(&message), 3u128.pow(1) as f64 / 8 as f64);
+    assert_eq!(compute_pow_score(&message), 3u128.pow(1) as f64 / 8_f64);
 
     let message: [u8; 8] = [203, 124, 2, 0, 0, 0, 0, 0];
 
-    assert_eq!(compute_pow_score(&message), 3u128.pow(10) as f64 / 8 as f64);
+    assert_eq!(compute_pow_score(&message), 3u128.pow(10) as f64 / 8_f64);
 
     let message: [u8; 8] = [65, 235, 119, 85, 85, 85, 85, 85];
 
-    assert_eq!(compute_pow_score(&message), 3u128.pow(14) as f64 / 8 as f64);
+    assert_eq!(compute_pow_score(&message), 3u128.pow(14) as f64 / 8_f64);
 
     let message: [u8; 10000] = [0; 10000];
 
-    assert_eq!(compute_pow_score(&message), 3u128.pow(0) as f64 / 10000 as f64);
+    assert_eq!(compute_pow_score(&message), 3u128.pow(0) as f64 / 10000_f64);
 }
